requirements.txt
accelerate==1.3.0
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==4.0.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.1
datasets==3.2.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.16.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
llama_cpp_python @ file:///C:/Users/misha/PycharmProjects/machinelearning/llama_cpp_python-0.3.4-cp310-cp310-win_amd64.whl#sha256=626804da610821a5c2c8865126637a49e6e6e5de34ac4e9a74234dcad20939e7
lxml==5.3.1
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
peft==0.14.0
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdfium2==4.30.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-build==0.18.1
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.51.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.2
xxhash==3.5.0
yarl==1.18.3
